feat(liquidation_process): Plumbing
