+++ src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod indexer;
+pub mod monitoring;
