+++ src/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod asset;
+pub mod position;
