+++ src/types/position.rs
@@ -1,39 +1,11 @@
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use apibara_core::starknet::v1alpha2::FieldElement;
 use bigdecimal::BigDecimal;
 use starknet::core::types::Felt;
 
 use crate::{
-    config::{get_asset_name_for_address, get_decimal_for_address},
-    conversions::apibara_field_element_as_felt,
-    oracle::PragmaOracle,
+    oracle::PragmaOracle, types::asset::Asset, utils::conversions::apibara_field_element_as_felt,
 };
-
-#[derive(Default, Clone, Hash, Eq, PartialEq, Debug)]
-pub struct Asset {
-    pub name: String,
-    pub address: Felt,
-    pub amount: BigDecimal,
-    pub decimals: i64,
-}
-
-impl TryFrom<Felt> for Asset {
-    type Error = anyhow::Error;
-
-    fn try_from(value: Felt) -> Result<Self> {
-        let name = get_asset_name_for_address(value)
-            .ok_or_else(|| anyhow!("Failed to get asset name for address"))?;
-        let decimals = get_decimal_for_address(value)
-            .ok_or_else(|| anyhow!("Failed to get decimals for address"))?;
-
-        Ok(Asset {
-            name,
-            address: value,
-            amount: BigDecimal::from(0),
-            decimals,
-        })
-    }
-}
 
 #[derive(Default, Clone, Hash, Eq, PartialEq, Debug)]
 pub struct Position {
@@ -91,31 +63,6 @@ impl Position {
             self.collateral.address,
             self.debt.address,
             self.user_address,
-        ]
-    }
-}
-
-#[derive(Default, Clone, Hash, Eq, PartialEq, Debug)]
-pub struct GetLTVConfigRequest {
-    pub pool_id: Felt,
-    pub collateral_asset_address: Felt,
-    pub debt_asset_address: Felt,
-}
-
-impl GetLTVConfigRequest {
-    pub fn try_from_event_keys(event_keys: &[Felt]) -> GetLTVConfigRequest {
-        GetLTVConfigRequest {
-            pool_id: event_keys[1],
-            collateral_asset_address: event_keys[2],
-            debt_asset_address: event_keys[3],
-        }
-    }
-
-    pub fn as_calldata(&self) -> Vec<Felt> {
-        vec![
-            self.pool_id,
-            self.collateral_asset_address,
-            self.debt_asset_address,
         ]
     }
 }
